# [cidr] Use a decision tree for faster slow search