## Context

The slow search for IP addresses is very slow, on random ip addresses it took around 0.1s / ip searched on my laptop.

This is because we compare the ip with every ip range in the lists individually, so the complexity is linear in the size of the warning list.

Instead, we can build a decision tree that take the bits from most significant to less significant. We stop when the path we take in the tree matches a range. This makes the search constant in time complexity (we need at most 32 comparison for ipv4, and 128 for ipv6)

For eg, if we consider the 2 ranges "192.0.0.0/2" (which equals to the mask `"11*"`) and "160.0.0.0/3" ("101*"), we can use the tree:
 - 0: False
 - 1: 
   - 0: 
     - 0: False
     - 1: True (matches 101*, which is  "160.0.0.0/3")
   - 1: True (matches 11*, which is "192.0.0.0/2")

## Results

Searching on 100 random ip addresses: 10 seconds => 0.01 seconds (0.1 seconds / ip => 0.1 milliseconds / ip)

Building all the warning lists: 1 seconds => 6 seconds (but we need to do it only one time)